ES6 rewrite
