+++ package.json
@@ -40,5 +40,10 @@
   },
   "peerDependencies": {
     "yo": ">=1.2.0"
+  },
+  "devDependencies": {
+    "babel-plugin-add-module-exports": "^0.1.2",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-register": "^6.3.13"
   }
 }
